# clawhub-main/public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenClaw API",
    "version": "1.0.0",
    "description": "Public REST API for skills. Rate limits: read 120/min per IP + 600/min per key; write 30/min per IP + 120/min per key."
  },
  "servers": [{ "url": "https://clawhub.ai" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token"
      }
    },
    "schemas": {
      "SearchResult": {
        "type": "object",
        "properties": {
          "score": { "type": "number" },
          "slug": { "type": ["string", "null"] },
          "displayName": { "type": ["string", "null"] },
          "summary": { "type": ["string", "null"] },
          "version": { "type": ["string", "null"] },
          "updatedAt": { "type": ["number", "null"] }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "results": { "type": "array", "items": { "$ref": "#/components/schemas/SearchResult" } }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/search": {
      "get": {
        "summary": "Search skills",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "highlightedOnly", "in": "query", "required": false, "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchResponse" } } }
          }
        }
      }
    },
...
  }
}